pyproject.toml
[project]
name = "telegrafono"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.10",
    "tomlkit>=0.13.2",
    "typer>=0.15.1",
]

[project.scripts]
telegrafono = "telegrafono.__main__:main"

[tool.uv]
package = true